requirements.txt
pySerial
configparser
numpy
pillow
